More changes for LDBLE
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/trunk@2795 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/basic.c
@@ -4129,11 +4129,11 @@ cmdpunch (struct LOC_exec *LINK)
     }
     else if (punch.high_precision == FALSE)
     {
-      fpunchf_user (n_user_punch_index, "%12.4e\t", n.UU.val);
+      fpunchf_user (n_user_punch_index, "%12.4e\t", (double) n.UU.val);
     }
     else
     {
-      fpunchf_user (n_user_punch_index, "%20.12e\t", n.UU.val);
+      fpunchf_user (n_user_punch_index, "%20.12e\t", (double) n.UU.val);
     }
     ++n_user_punch_index;
   }
